refactor(app): hide implementations details of initialization of the RecyclerView to AbstractListActivity
